(server) refactoring : service와 controller 분리
