+++ my-diary-back/my-diary-back/src/main/java/kr/ac/kumoh/oiyo/mydiaryback/service/UserService.java
@@ -1,0 +1,4 @@
+package kr.ac.kumoh.oiyo.mydiaryback.service;
+
+public class UserService {
+}
